+++ internal/plugins/gitstatus/plugin.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"strings"
 	"time"
 
 	"github.com/charmbracelet/bubbles/textarea"
@@ -181,19 +180,6 @@ func New() *Plugin {
 	}
 }
 
-// getHighlighter returns a syntax highlighter for the given filename.
-// Caches the highlighter to avoid re-creating on every render.
-func (p *Plugin) getHighlighter(filename string) *SyntaxHighlighter {
-	if filename == "" {
-		return nil
-	}
-	if p.syntaxHighlighterFile == filename && p.syntaxHighlighter != nil {
-		return p.syntaxHighlighter
-	}
-	p.syntaxHighlighter = NewSyntaxHighlighter(filename)
-	p.syntaxHighlighterFile = filename
-	return p.syntaxHighlighter
-}
 
 // ID returns the plugin identifier.
 func (p *Plugin) ID() string { return pluginID }
@@ -589,663 +575,7 @@ func (p *Plugin) Update(msg tea.Msg) (plugin.Plugin, tea.Cmd) {
 	return p, nil
 }
 
-// totalSelectableItems returns the count of all selectable items (files + commits).
-func (p *Plugin) totalSelectableItems() int {
-	entries := p.tree.AllEntries()
-	commits := p.recentCommits
-	if p.historyFilterActive && p.filteredCommits != nil {
-		commits = p.filteredCommits
-	}
-	return len(entries) + len(commits)
-}
 
-// cursorOnCommit returns true if cursor is on a commit (past all files).
-func (p *Plugin) cursorOnCommit() bool {
-	return p.cursor >= len(p.tree.AllEntries())
-}
-
-// activeCommits returns filtered commits if filter active, otherwise recent commits.
-func (p *Plugin) activeCommits() []*Commit {
-	if p.historyFilterActive && p.filteredCommits != nil {
-		return p.filteredCommits
-	}
-	return p.recentCommits
-}
-
-// selectedCommitIndex returns the index into active commits when cursor is on commit.
-func (p *Plugin) selectedCommitIndex() int {
-	entries := p.tree.AllEntries()
-	return p.cursor - len(entries)
-}
-
-func (p *Plugin) toggleSidebar() {
-	if p.sidebarVisible {
-		p.sidebarRestore = p.activePane
-		p.sidebarVisible = false
-		if p.activePane == PaneSidebar {
-			p.activePane = PaneDiff
-		}
-		return
-	}
-
-	p.sidebarVisible = true
-	if p.sidebarRestore == PaneSidebar {
-		p.activePane = PaneSidebar
-	} else {
-		p.activePane = PaneDiff
-	}
-}
-
-// updateStatus handles key events in the status view.
-func (p *Plugin) updateStatus(msg tea.KeyMsg) (plugin.Plugin, tea.Cmd) {
-	// Handle diff pane keys when focused on diff
-	if p.activePane == PaneDiff {
-		return p.updateStatusDiffPane(msg)
-	}
-
-	entries := p.tree.AllEntries()
-	totalItems := p.totalSelectableItems()
-
-	switch msg.String() {
-	case "j", "down":
-		if p.cursor < totalItems-1 {
-			p.cursor++
-			p.ensureCursorVisible()
-			if p.cursorOnCommit() {
-				commitIdx := p.selectedCommitIndex()
-				p.ensureCommitVisible(commitIdx)
-				// Trigger load-more when within 3 commits of end (only for unfiltered view)
-				var loadMoreCmd tea.Cmd
-				commits := p.activeCommits()
-				if !p.historyFilterActive && p.moreCommitsAvailable && commitIdx >= len(commits)-3 && !p.loadingMoreCommits {
-					loadMoreCmd = p.loadMoreCommits()
-				}
-				return p, tea.Batch(p.autoLoadCommitPreview(), loadMoreCmd)
-			}
-			return p, p.autoLoadDiff()
-		}
-		return p, nil
-
-	case "k", "up":
-		if p.cursor > 0 {
-			p.cursor--
-			p.ensureCursorVisible()
-			if p.cursorOnCommit() {
-				p.ensureCommitVisible(p.selectedCommitIndex())
-				return p, p.autoLoadCommitPreview()
-			}
-			return p, p.autoLoadDiff()
-		}
-		return p, nil
-
-	case "g":
-		p.cursor = 0
-		p.scrollOff = 0
-		p.commitScrollOff = 0 // Reset commit scroll when jumping to top
-		if p.cursorOnCommit() {
-			return p, p.autoLoadCommitPreview()
-		}
-		return p, p.autoLoadDiff()
-
-	case "G":
-		if totalItems > 0 {
-			p.cursor = totalItems - 1
-			p.ensureCursorVisible()
-			if p.cursorOnCommit() {
-				commitIdx := p.selectedCommitIndex()
-				p.ensureCommitVisible(commitIdx)
-				// Trigger load-more when jumping to end
-				var loadMoreCmd tea.Cmd
-				if p.moreCommitsAvailable && commitIdx >= len(p.recentCommits)-3 && !p.loadingMoreCommits {
-					loadMoreCmd = p.loadMoreCommits()
-				}
-				return p, tea.Batch(p.autoLoadCommitPreview(), loadMoreCmd)
-			}
-			return p, p.autoLoadDiff()
-		}
-		return p, nil
-
-	case "l", "right":
-		// Focus diff pane (when on a file) or commit preview pane (when on a commit)
-		if p.sidebarVisible {
-			if p.cursorOnCommit() && p.previewCommit != nil {
-				p.activePane = PaneDiff
-			} else if p.selectedDiffFile != "" {
-				p.activePane = PaneDiff
-			}
-		}
-
-	case "tab", "shift+tab":
-		// Switch focus to diff pane (if sidebar visible)
-		if p.sidebarVisible && (p.selectedDiffFile != "" || p.previewCommit != nil) {
-			p.activePane = PaneDiff
-		}
-
-	case "\\":
-		// Toggle sidebar visibility
-		p.toggleSidebar()
-
-	case "s":
-		if len(entries) > 0 && p.cursor < len(entries) {
-			entry := entries[p.cursor]
-			if !entry.Staged {
-				stagedCount := len(p.tree.Staged)
-				totalEntries := len(entries)
-
-				// Handle folder entries - stage all children
-				if entry.IsFolder {
-					for _, child := range entry.Children {
-						_ = p.tree.StageFile(child.Path)
-					}
-				} else {
-					if err := p.tree.StageFile(entry.Path); err != nil {
-						return p, nil
-					}
-				}
-				// After staging, move cursor to first unstaged file position
-				newFirstUnstaged := stagedCount + 1
-				if newFirstUnstaged < totalEntries {
-					p.cursor = newFirstUnstaged
-				} else {
-					p.cursor = totalEntries - 1
-				}
-				return p, tea.Batch(p.refresh(), p.loadRecentCommits())
-			}
-		}
-
-	case "u":
-		if len(entries) > 0 && p.cursor < len(entries) {
-			entry := entries[p.cursor]
-			if entry.Staged {
-				if err := p.tree.UnstageFile(entry.Path); err == nil {
-					return p, tea.Batch(p.refresh(), p.loadRecentCommits())
-				}
-			}
-		}
-
-	case "d":
-		// Open full-screen diff view for files
-		if !p.cursorOnCommit() && len(entries) > 0 && p.cursor < len(entries) {
-			entry := entries[p.cursor]
-			p.diffReturnMode = p.viewMode
-			p.viewMode = ViewModeDiff
-			p.diffFile = entry.Path
-			p.diffCommit = ""
-			p.diffScroll = 0
-			if entry.IsFolder {
-				return p, p.loadFullFolderDiff(entry)
-			}
-			return p, p.loadDiff(entry.Path, entry.Staged, entry.Status)
-		}
-		// For commits, focus the preview pane (same as l/right)
-		if p.cursorOnCommit() && p.previewCommit != nil {
-			p.activePane = PaneDiff
-		}
-
-	case "enter":
-		// For folders: toggle expand/collapse
-		// For files: open in editor
-		// For commits: focus the preview pane
-		if p.cursorOnCommit() {
-			if p.previewCommit != nil {
-				p.activePane = PaneDiff
-			}
-		} else if len(entries) > 0 && p.cursor < len(entries) {
-			entry := entries[p.cursor]
-			if entry.IsFolder {
-				// Toggle folder expansion
-				entry.IsExpanded = !entry.IsExpanded
-				// Reload diff for this folder
-				return p, p.autoLoadDiff()
-			}
-			return p, p.openFile(entry.Path)
-		}
-
-	case "r":
-		p.pushError = "" // Clear any stale push error
-		return p, tea.Batch(p.refresh(), p.loadRecentCommits())
-
-	case "S":
-		// Stage all files
-		if err := p.tree.StageAll(); err == nil {
-			return p, tea.Batch(p.refresh(), p.loadRecentCommits())
-		}
-
-	case "O":
-		// Open file in file browser (for files only, not commits)
-		if !p.cursorOnCommit() && len(entries) > 0 && p.cursor < len(entries) {
-			entry := entries[p.cursor]
-			return p, p.openInFileBrowser(entry.Path)
-		}
-
-	case "c":
-		// Enter commit mode only if staged files exist
-		if p.tree.HasStagedFiles() {
-			p.viewMode = ViewModeCommit
-			p.initCommitTextarea()
-			return p, nil
-		}
-
-	case "P":
-		// Open push menu (following lazygit convention)
-		if p.canPush() && !p.pushInProgress {
-			p.pushMenuReturnMode = p.viewMode
-			p.viewMode = ViewModePushMenu
-		}
-
-	case "y":
-		// Yank commit as markdown (when on commit in sidebar)
-		if p.cursorOnCommit() {
-			return p, p.copyCommitToClipboard()
-		}
-
-	case "Y":
-		// Yank commit ID (when on commit in sidebar)
-		if p.cursorOnCommit() {
-			return p, p.copyCommitIDToClipboard()
-		}
-
-	case "o":
-		// Open commit in GitHub (when on commit in sidebar)
-		if p.cursorOnCommit() {
-			return p, p.openCommitInGitHub()
-		}
-
-	case "D":
-		// Discard changes (confirm modal) - only for modified/staged files, not commits
-		if !p.cursorOnCommit() && len(entries) > 0 && p.cursor < len(entries) {
-			entry := entries[p.cursor]
-			// Don't allow discard on untracked folders (would delete)
-			if entry.IsFolder && entry.Status == StatusUntracked {
-				return p, nil
-			}
-			p.discardFile = entry
-			p.discardReturnMode = p.viewMode
-			p.viewMode = ViewModeConfirmDiscard
-		}
-
-	case "z":
-		// Stash current changes (if there are any)
-		if p.tree.TotalCount() > 0 {
-			return p, p.doStashPush()
-		}
-
-	case "Z":
-		// Pop latest stash - show confirm modal first
-		return p, p.confirmStashPop()
-
-	case "b":
-		// Open branch picker
-		p.branchReturnMode = p.viewMode
-		p.branchCursor = 0
-		p.viewMode = ViewModeBranchPicker
-		return p, p.loadBranches()
-
-	case "f":
-		// On commit: filter by author; on file: fetch
-		if p.cursorOnCommit() {
-			commits := p.activeCommits()
-			commitIdx := p.selectedCommitIndex()
-			if commitIdx >= 0 && commitIdx < len(commits) {
-				commit := commits[commitIdx]
-				p.historyFilterAuthor = commit.Author
-				p.historyFilterActive = true
-				return p, p.loadFilteredCommits()
-			}
-		} else {
-			// Fetch from remote
-			if !p.fetchInProgress {
-				p.fetchInProgress = true
-				p.fetchError = ""
-				p.fetchSuccess = false
-				return p, p.doFetch()
-			}
-		}
-
-	case "F":
-		// Clear all history filters
-		if p.historyFilterActive {
-			p.historyFilterAuthor = ""
-			p.historyFilterPath = ""
-			p.historyFilterActive = false
-			p.filteredCommits = nil
-			// Recompute graph for unfiltered commits
-			if p.showCommitGraph && len(p.recentCommits) > 0 {
-				p.commitGraphLines = ComputeGraphForCommits(p.recentCommits)
-			}
-		}
-
-	case "p":
-		// On commit: filter by path (open modal); on file: pull
-		if p.cursorOnCommit() {
-			p.pathFilterMode = true
-			p.pathFilterInput = ""
-			return p, nil
-		}
-		// Pull from remote (only if no local changes to avoid conflicts)
-		if !p.pullInProgress {
-			p.pullInProgress = true
-			p.pullError = ""
-			p.pullSuccess = false
-			return p, p.doPull()
-		}
-
-	case "/":
-		// Open history search modal (only when on commits)
-		if p.cursorOnCommit() {
-			if p.historySearchState == nil {
-				p.historySearchState = NewHistorySearchState()
-			}
-			p.historySearchState.Reset()
-			p.historySearchMode = true
-			return p, nil
-		}
-
-	case "n":
-		// Next search match (after search committed)
-		if p.historySearchState != nil && p.historySearchState.Committed && len(p.historySearchState.Matches) > 0 {
-			p.historySearchState.Cursor++
-			if p.historySearchState.Cursor >= len(p.historySearchState.Matches) {
-				p.historySearchState.Cursor = 0 // Wrap around
-			}
-			return p, p.jumpToSearchMatch()
-		}
-
-	case "N":
-		// Previous search match
-		if p.historySearchState != nil && p.historySearchState.Committed && len(p.historySearchState.Matches) > 0 {
-			p.historySearchState.Cursor--
-			if p.historySearchState.Cursor < 0 {
-				p.historySearchState.Cursor = len(p.historySearchState.Matches) - 1 // Wrap around
-			}
-			return p, p.jumpToSearchMatch()
-		}
-
-	case "esc":
-		// ESC clears search state (if any active search)
-		if p.historySearchState != nil && p.historySearchState.Committed {
-			p.clearSearchState()
-			return p, nil
-		}
-
-	case "v":
-		// Toggle commit graph display (only when on commits)
-		if p.cursorOnCommit() {
-			p.showCommitGraph = !p.showCommitGraph
-			_ = state.SetGitGraphEnabled(p.showCommitGraph)
-			if p.showCommitGraph {
-				commits := p.activeCommits()
-				p.commitGraphLines = ComputeGraphForCommits(commits)
-			}
-		}
-		return p, nil
-	}
-
-	return p, nil
-}
-
-// updateStatusDiffPane handles key events when the diff pane is focused.
-func (p *Plugin) updateStatusDiffPane(msg tea.KeyMsg) (plugin.Plugin, tea.Cmd) {
-	// If showing commit preview, handle file list navigation
-	if p.previewCommit != nil && p.cursorOnCommit() {
-		return p.updateCommitPreviewPane(msg)
-	}
-
-	switch msg.String() {
-	case "esc":
-		// ESC always returns to sidebar
-		p.activePane = PaneSidebar
-
-	case "h", "left":
-		// Horizontal scroll left (use ESC or Tab to switch panes)
-		if p.diffPaneHorizScroll > 0 {
-			p.diffPaneHorizScroll -= 10
-			if p.diffPaneHorizScroll < 0 {
-				p.diffPaneHorizScroll = 0
-			}
-		}
-
-	case "l", "right":
-		// Horizontal scroll right
-		p.diffPaneHorizScroll += 10
-		p.clampDiffPaneHorizScroll()
-
-	case "j", "down":
-		p.diffPaneScroll++
-
-	case "k", "up":
-		if p.diffPaneScroll > 0 {
-			p.diffPaneScroll--
-		}
-
-	case "g":
-		p.diffPaneScroll = 0
-		p.diffPaneHorizScroll = 0
-
-	case "G":
-		if p.diffPaneParsedDiff != nil {
-			lines := countParsedDiffLines(p.diffPaneParsedDiff)
-			maxScroll := lines - (p.height - 6)
-			if maxScroll > 0 {
-				p.diffPaneScroll = maxScroll
-			}
-		}
-
-	case "ctrl+d":
-		p.diffPaneScroll += 10
-		// Clamp to max
-		if p.diffPaneParsedDiff != nil {
-			lines := countParsedDiffLines(p.diffPaneParsedDiff)
-			maxScroll := lines - (p.height - 6)
-			if maxScroll < 0 {
-				maxScroll = 0
-			}
-			if p.diffPaneScroll > maxScroll {
-				p.diffPaneScroll = maxScroll
-			}
-		}
-
-	case "ctrl+u":
-		p.diffPaneScroll -= 10
-		if p.diffPaneScroll < 0 {
-			p.diffPaneScroll = 0
-		}
-
-	case "0":
-		// Reset horizontal scroll
-		p.diffPaneHorizScroll = 0
-
-	case "v":
-		// Toggle view mode (unified/side-by-side) for inline diff pane
-		if p.diffPaneViewMode == DiffViewUnified {
-			p.diffPaneViewMode = DiffViewSideBySide
-		} else {
-			p.diffPaneViewMode = DiffViewUnified
-		}
-
-	case "tab", "shift+tab":
-		// Switch focus to sidebar (if visible)
-		if p.sidebarVisible {
-			p.activePane = PaneSidebar
-		}
-
-	case "\\":
-		// Toggle sidebar visibility
-		p.toggleSidebar()
-
-	case "d":
-		// Open full-screen diff view for current file
-		entries := p.tree.AllEntries()
-		if len(entries) > 0 && p.cursor < len(entries) {
-			entry := entries[p.cursor]
-			p.diffReturnMode = p.viewMode
-			p.viewMode = ViewModeDiff
-			p.diffFile = entry.Path
-			p.diffCommit = ""
-			p.diffScroll = 0
-			return p, p.loadDiff(entry.Path, entry.Staged, entry.Status)
-		}
-	}
-
-	return p, nil
-}
-
-// updateCommitPreviewPane handles key events when viewing commit preview in the diff pane.
-func (p *Plugin) updateCommitPreviewPane(msg tea.KeyMsg) (plugin.Plugin, tea.Cmd) {
-	c := p.previewCommit
-	if c == nil {
-		return p, nil
-	}
-
-	switch msg.String() {
-	case "esc", "h", "left":
-		// Return to sidebar
-		p.activePane = PaneSidebar
-
-	case "j", "down":
-		// Navigate file list
-		if p.previewCommitCursor < len(c.Files)-1 {
-			p.previewCommitCursor++
-			p.ensurePreviewCursorVisible()
-		}
-
-	case "k", "up":
-		if p.previewCommitCursor > 0 {
-			p.previewCommitCursor--
-			p.ensurePreviewCursorVisible()
-		}
-
-	case "g":
-		p.previewCommitCursor = 0
-		p.previewCommitScroll = 0
-
-	case "G":
-		if len(c.Files) > 0 {
-			p.previewCommitCursor = len(c.Files) - 1
-			p.ensurePreviewCursorVisible()
-		}
-
-	case "enter", "d":
-		// Open full-screen diff for selected file in commit
-		if p.previewCommitCursor < len(c.Files) {
-			file := c.Files[p.previewCommitCursor]
-			p.diffReturnMode = p.viewMode
-			p.viewMode = ViewModeDiff
-			p.diffFile = file.Path
-			p.diffCommit = c.Hash
-			p.diffScroll = 0
-			return p, p.loadCommitFileDiff(c.Hash, file.Path)
-		}
-
-	case "tab", "shift+tab":
-		// Switch focus to sidebar (if visible)
-		if p.sidebarVisible {
-			p.activePane = PaneSidebar
-		}
-
-	case "\\":
-		// Toggle sidebar visibility
-		p.toggleSidebar()
-
-	case "y":
-		// Yank commit as markdown
-		return p, p.copyCommitToClipboard()
-
-	case "Y":
-		// Yank commit ID
-		return p, p.copyCommitIDToClipboard()
-
-	case "o":
-		// Open commit in GitHub
-		return p, p.openCommitInGitHub()
-	}
-
-	return p, nil
-}
-
-// ensurePreviewCursorVisible adjusts scroll to keep commit preview cursor visible.
-func (p *Plugin) ensurePreviewCursorVisible() {
-	// Estimate visible file rows (rough - matches renderCommitPreview calculation)
-	visibleRows := p.height - 15
-	if visibleRows < 3 {
-		visibleRows = 3
-	}
-
-	if p.previewCommitCursor < p.previewCommitScroll {
-		p.previewCommitScroll = p.previewCommitCursor
-	} else if p.previewCommitCursor >= p.previewCommitScroll+visibleRows {
-		p.previewCommitScroll = p.previewCommitCursor - visibleRows + 1
-	}
-}
-
-// autoLoadDiff triggers loading the diff for the currently selected file or folder.
-func (p *Plugin) autoLoadDiff() tea.Cmd {
-	entries := p.tree.AllEntries()
-	if len(entries) == 0 || p.cursor >= len(entries) {
-		p.selectedDiffFile = ""
-		p.diffPaneParsedDiff = nil
-		return nil
-	}
-
-	entry := entries[p.cursor]
-	if entry.Path == p.selectedDiffFile {
-		return nil // Already loaded
-	}
-
-	p.selectedDiffFile = entry.Path
-	// Keep old diff visible until new one loads (prevents flashing)
-	p.diffPaneScroll = 0
-	// Clear commit preview when switching to file
-	p.previewCommit = nil
-
-	// Handle folder entries
-	if entry.IsFolder {
-		return p.loadFolderDiff(entry)
-	}
-
-	return p.loadInlineDiff(entry.Path, entry.Staged, entry.Status)
-}
-
-// autoLoadCommitPreview triggers loading commit detail for the currently selected commit.
-func (p *Plugin) autoLoadCommitPreview() tea.Cmd {
-	if !p.cursorOnCommit() {
-		p.previewCommit = nil
-		return nil
-	}
-
-	commits := p.activeCommits()
-	commitIdx := p.selectedCommitIndex()
-	if commitIdx < 0 || commitIdx >= len(commits) {
-		p.previewCommit = nil
-		return nil
-	}
-
-	commit := commits[commitIdx]
-	// Already loaded this commit?
-	if p.previewCommit != nil && p.previewCommit.Hash == commit.Hash {
-		return nil
-	}
-
-	// Clear file diff when switching to commit
-	p.selectedDiffFile = ""
-	p.diffPaneParsedDiff = nil
-	p.previewCommitCursor = 0
-	p.previewCommitScroll = 0
-
-	return p.loadCommitDetailForPreview(commit.Hash)
-}
-
-// loadCommitDetailForPreview loads commit detail for inline preview.
-func (p *Plugin) loadCommitDetailForPreview(hash string) tea.Cmd {
-	workDir := p.ctx.WorkDir
-	return func() tea.Msg {
-		commit, err := GetCommitDetail(workDir, hash)
-		if err != nil {
-			return ErrorMsg{Err: err}
-		}
-		return CommitPreviewLoadedMsg{Commit: commit}
-	}
-}
 
 // CommitPreviewLoadedMsg is sent when commit preview is loaded.
 type CommitPreviewLoadedMsg struct {
@@ -1264,100 +594,6 @@ func countParsedDiffLines(diff *ParsedDiff) int {
 	return count
 }
 
-// updateDiff handles key events in the diff view.
-func (p *Plugin) updateDiff(msg tea.KeyMsg) (plugin.Plugin, tea.Cmd) {
-	switch msg.String() {
-	case "esc", "q":
-		// Return to previous view
-		p.diffContent = ""
-		p.diffRaw = ""
-		p.parsedDiff = nil
-		p.diffHorizOff = 0
-		p.diffCommit = ""
-		p.diffFile = ""
-		p.viewMode = p.diffReturnMode
-		// If returning to status view with commit preview, focus the preview pane
-		if p.diffReturnMode == ViewModeStatus && p.previewCommit != nil {
-			p.activePane = PaneDiff
-		}
-
-	case "j", "down":
-		p.diffScroll++
-
-	case "k", "up":
-		if p.diffScroll > 0 {
-			p.diffScroll--
-		}
-
-	case "g":
-		p.diffScroll = 0
-		p.diffHorizOff = 0
-
-	case "G":
-		lines := countLines(p.diffContent)
-		maxScroll := lines - (p.height - 2)
-		if maxScroll > 0 {
-			p.diffScroll = maxScroll
-		}
-
-	case "v":
-		// Toggle between unified and side-by-side view
-		if p.diffViewMode == DiffViewUnified {
-			p.diffViewMode = DiffViewSideBySide
-			_ = state.SetGitDiffMode("side-by-side")
-		} else {
-			p.diffViewMode = DiffViewUnified
-			_ = state.SetGitDiffMode("unified")
-		}
-		p.diffHorizOff = 0
-
-	case "\\":
-		// Toggle sidebar visibility
-		p.toggleSidebar()
-
-	case "h", "left", "<", "H":
-		// Horizontal scroll left
-		if p.diffHorizOff > 0 {
-			p.diffHorizOff -= 10
-			if p.diffHorizOff < 0 {
-				p.diffHorizOff = 0
-			}
-		}
-
-	case "l", "right", ">", "L":
-		// Horizontal scroll right
-		p.diffHorizOff += 10
-		p.clampDiffHorizScroll()
-
-	case "ctrl+d":
-		// Page down (~10 lines)
-		p.diffScroll += 10
-		// Clamp to max
-		lines := countLines(p.diffContent)
-		maxScroll := lines - (p.height - 2)
-		if maxScroll < 0 {
-			maxScroll = 0
-		}
-		if p.diffScroll > maxScroll {
-			p.diffScroll = maxScroll
-		}
-
-	case "ctrl+u":
-		// Page up (~10 lines)
-		p.diffScroll -= 10
-		if p.diffScroll < 0 {
-			p.diffScroll = 0
-		}
-
-	case "O":
-		// Open file in file browser
-		if p.diffFile != "" {
-			return p, p.openInFileBrowser(p.diffFile)
-		}
-	}
-
-	return p, nil
-}
 
 // View renders the plugin.
 func (p *Plugin) View(width, height int) string {
@@ -1539,26 +775,6 @@ func (p *Plugin) listenForWatchEvents() tea.Cmd {
 	}
 }
 
-// loadDiff loads the diff for a file.
-func (p *Plugin) loadDiff(path string, staged bool, status FileStatus) tea.Cmd {
-	workDir := p.ctx.WorkDir
-	return func() tea.Msg {
-		var rawDiff string
-		var err error
-
-		// Untracked files need special handling - create new file diff
-		if status == StatusUntracked {
-			rawDiff, err = GetNewFileDiff(workDir, path)
-		} else {
-			rawDiff, err = GetDiff(workDir, path, staged)
-		}
-		if err != nil {
-			return ErrorMsg{Err: err}
-		}
-
-		return DiffLoadedMsg{Content: rawDiff, Raw: rawDiff}
-	}
-}
 
 // openFile opens a file in the default editor.
 func (p *Plugin) openFile(path string) tea.Cmd {
@@ -1582,35 +798,6 @@ func (p *Plugin) openInFileBrowser(path string) tea.Cmd {
 	)
 }
 
-// ensureCursorVisible adjusts scroll to keep cursor visible.
-func (p *Plugin) ensureCursorVisible() {
-	visibleRows := p.height - 4 // Account for header and section spacing
-	if visibleRows < 1 {
-		visibleRows = 1
-	}
-
-	if p.cursor < p.scrollOff {
-		p.scrollOff = p.cursor
-	} else if p.cursor >= p.scrollOff+visibleRows {
-		p.scrollOff = p.cursor - visibleRows + 1
-	}
-}
-
-// visibleCommitCount returns how many commits can display in the sidebar.
-func (p *Plugin) visibleCommitCount() int {
-	// paneHeight = p.height - 2 (borders)
-	// innerHeight = paneHeight - 2 (header)
-	paneHeight := p.height - 2
-	if paneHeight < 4 {
-		paneHeight = 4
-	}
-	innerHeight := paneHeight - 2
-	if innerHeight < 1 {
-		innerHeight = 1
-	}
-
-	return p.commitSectionCapacity(innerHeight)
-}
 
 func mergeRecentCommits(existing, latest []*Commit) []*Commit {
 	if len(latest) == 0 {
@@ -1650,159 +837,6 @@ func indexOfCommitHash(commits []*Commit, hash string) int {
 	return -1
 }
 
-func (p *Plugin) clampCommitScroll() {
-	visibleCommits := p.visibleCommitCount()
-	if visibleCommits < 1 {
-		visibleCommits = 1
-	}
-	maxOff := len(p.recentCommits) - visibleCommits
-	if maxOff < 0 {
-		maxOff = 0
-	}
-	if p.commitScrollOff < 0 {
-		p.commitScrollOff = 0
-	}
-	if p.commitScrollOff > maxOff {
-		p.commitScrollOff = maxOff
-	}
-}
-
-func (p *Plugin) ensureCommitListFilled() tea.Cmd {
-	if p.historyFilterActive || p.loadingMoreCommits || !p.moreCommitsAvailable {
-		return nil
-	}
-	visibleCommits := p.visibleCommitCount()
-	if visibleCommits < 1 || len(p.recentCommits) >= visibleCommits {
-		return nil
-	}
-	return p.loadMoreCommits()
-}
-
-// commitSectionCapacity returns how many commits can display in the sidebar
-// given the current layout and inner height.
-func (p *Plugin) commitSectionCapacity(visibleHeight int) int {
-	// visibleHeight already excludes the "Files" header lines.
-	linesUsed := 0
-
-	entries := p.tree.AllEntries()
-	if len(entries) == 0 {
-		// "Working tree clean"
-		linesUsed++
-	} else {
-		// Reserve space for commits when files are present (match renderSidebar)
-		commitsReserve := 5
-		if len(p.recentCommits) > 3 {
-			commitsReserve = 6
-		}
-		filesHeight := visibleHeight - commitsReserve - 2 // -2 for section headers
-		if filesHeight < 3 {
-			filesHeight = 3
-		}
-
-		lineNum := 0
-		if len(p.tree.Staged) > 0 && lineNum < filesHeight {
-			lineNum++ // section header
-			for range p.tree.Staged {
-				if lineNum >= filesHeight {
-					break
-				}
-				lineNum++
-			}
-		}
-		if len(p.tree.Modified) > 0 && lineNum < filesHeight {
-			if len(p.tree.Staged) > 0 {
-				if lineNum < filesHeight {
-					lineNum++ // blank line between sections
-				}
-			}
-			if lineNum < filesHeight {
-				lineNum++ // section header
-			}
-			for range p.tree.Modified {
-				if lineNum >= filesHeight {
-					break
-				}
-				lineNum++
-			}
-		}
-		if len(p.tree.Untracked) > 0 && lineNum < filesHeight {
-			if len(p.tree.Staged) > 0 || len(p.tree.Modified) > 0 {
-				if lineNum < filesHeight {
-					lineNum++ // blank line between sections
-				}
-			}
-			if lineNum < filesHeight {
-				lineNum++ // section header
-			}
-			for range p.tree.Untracked {
-				if lineNum >= filesHeight {
-					break
-				}
-				lineNum++
-			}
-		}
-		linesUsed += lineNum
-	}
-
-	// Separator (blank line + separator line)
-	linesUsed += 2
-
-	// Remote operation status line if present
-	if p.pushInProgress || p.fetchInProgress || p.pullInProgress ||
-		p.pushSuccess || p.fetchSuccess || p.pullSuccess ||
-		p.pushError != "" || p.fetchError != "" || p.pullError != "" {
-		linesUsed++
-	}
-
-	// Commit header
-	linesUsed += 1
-
-	available := visibleHeight - linesUsed
-	if available < 2 {
-		available = 2
-	}
-	return available
-}
-
-// ensureCommitVisible adjusts commitScrollOff to keep selected commit visible.
-// commitIdx is the absolute index into activeCommits.
-func (p *Plugin) ensureCommitVisible(commitIdx int) {
-	commits := p.activeCommits()
-	if len(commits) == 0 {
-		p.commitScrollOff = 0
-		return
-	}
-
-	// Use a conservative estimate for visible commits
-	visibleCommits := p.visibleCommitCount()
-	if visibleCommits < 1 {
-		visibleCommits = 1
-	}
-
-	// Only adjust scroll if absolutely necessary
-	if commitIdx < p.commitScrollOff {
-		// Commit is above visible area - scroll up to show it at top
-		p.commitScrollOff = commitIdx
-	} else if commitIdx >= p.commitScrollOff+visibleCommits {
-		// Commit is below visible area - scroll down minimally
-		p.commitScrollOff = commitIdx - visibleCommits + 1
-	}
-	// If commit is within visible range, don't adjust scroll at all
-
-	// Clamp scroll offset to valid range
-	if p.commitScrollOff < 0 {
-		p.commitScrollOff = 0
-	}
-	// Ensure we don't scroll past where we'd have empty rows
-	// This is the key: maxOff ensures the visible area is always filled
-	maxOff := len(commits) - visibleCommits
-	if maxOff < 0 {
-		maxOff = 0
-	}
-	if p.commitScrollOff > maxOff {
-		p.commitScrollOff = maxOff
-	}
-}
 
 // countLines counts newlines in a string.
 func countLines(s string) int {
@@ -1938,128 +972,6 @@ type FetchSuccessClearMsg struct{}
 // PullSuccessClearMsg is sent to clear the pull success indicator.
 type PullSuccessClearMsg struct{}
 
-// loadInlineDiff loads a diff for inline preview in the three-pane view.
-func (p *Plugin) loadInlineDiff(path string, staged bool, status FileStatus) tea.Cmd {
-	workDir := p.ctx.WorkDir
-	return func() tea.Msg {
-		var rawDiff string
-		var err error
-
-		// Untracked files need special handling - create new file diff
-		if status == StatusUntracked {
-			rawDiff, err = GetNewFileDiff(workDir, path)
-		} else {
-			rawDiff, err = GetDiff(workDir, path, staged)
-		}
-		if err != nil {
-			return InlineDiffLoadedMsg{File: path, Raw: "", Parsed: nil}
-		}
-		parsed, _ := ParseUnifiedDiff(rawDiff)
-		return InlineDiffLoadedMsg{File: path, Raw: rawDiff, Parsed: parsed}
-	}
-}
-
-// loadRecentCommits loads recent commits for the sidebar with push status.
-func (p *Plugin) loadRecentCommits() tea.Cmd {
-	workDir := p.ctx.WorkDir
-	return func() tea.Msg {
-		commits, pushStatus, err := GetCommitHistoryWithPushStatus(workDir, commitHistoryPageSize)
-		if err != nil {
-			return RecentCommitsLoadedMsg{Commits: nil, PushStatus: nil}
-		}
-		return RecentCommitsLoadedMsg{Commits: commits, PushStatus: pushStatus}
-	}
-}
-
-// loadMoreCommits fetches the next batch of commits for infinite scroll.
-func (p *Plugin) loadMoreCommits() tea.Cmd {
-	if p.loadingMoreCommits {
-		return nil
-	}
-	p.loadingMoreCommits = true
-
-	workDir := p.ctx.WorkDir
-	skip := len(p.recentCommits)
-	return func() tea.Msg {
-		commits, pushStatus, err := GetCommitHistoryWithPushStatusOffset(workDir, commitHistoryPageSize, skip)
-		if err != nil {
-			return MoreCommitsLoadedMsg{Commits: nil, PushStatus: nil}
-		}
-		return MoreCommitsLoadedMsg{Commits: commits, PushStatus: pushStatus}
-	}
-}
-
-// loadCommitStats fetches stats for a specific commit (lazy loading).
-func (p *Plugin) loadCommitStats(hash string) tea.Cmd {
-	workDir := p.ctx.WorkDir
-	return func() tea.Msg {
-		commit, err := GetCommitDetail(workDir, hash)
-		if err != nil || commit == nil {
-			return CommitStatsLoadedMsg{Hash: hash, Stats: CommitStats{}}
-		}
-		return CommitStatsLoadedMsg{Hash: hash, Stats: commit.Stats}
-	}
-}
-
-// loadFilteredCommits fetches commits with current filter options.
-func (p *Plugin) loadFilteredCommits() tea.Cmd {
-	workDir := p.ctx.WorkDir
-	opts := HistoryFilterOpts{
-		Author: p.historyFilterAuthor,
-		Path:   p.historyFilterPath,
-		Limit:  50,
-	}
-	return func() tea.Msg {
-		commits, pushStatus, err := GetCommitHistoryFilteredWithPushStatus(workDir, opts)
-		if err != nil {
-			return FilteredCommitsLoadedMsg{Commits: nil, PushStatus: nil}
-		}
-		return FilteredCommitsLoadedMsg{Commits: commits, PushStatus: pushStatus}
-	}
-}
-
-// loadFolderDiff loads a concatenated diff for all files in a folder.
-func (p *Plugin) loadFolderDiff(entry *FileEntry) tea.Cmd {
-	workDir := p.ctx.WorkDir
-	folderPath := entry.Path
-	children := entry.Children
-	return func() tea.Msg {
-		rawDiff, err := GetFolderDiff(workDir, children)
-		if err != nil {
-			return InlineDiffLoadedMsg{File: folderPath, Raw: "", Parsed: nil}
-		}
-		parsed, _ := ParseUnifiedDiff(rawDiff)
-		return InlineDiffLoadedMsg{File: folderPath, Raw: rawDiff, Parsed: parsed}
-	}
-}
-
-// loadFullFolderDiff loads a concatenated diff for full-screen view.
-func (p *Plugin) loadFullFolderDiff(entry *FileEntry) tea.Cmd {
-	workDir := p.ctx.WorkDir
-	children := entry.Children
-	return func() tea.Msg {
-		rawDiff, err := GetFolderDiff(workDir, children)
-		if err != nil {
-			return ErrorMsg{Err: err}
-		}
-
-		return DiffLoadedMsg{Content: rawDiff, Raw: rawDiff}
-	}
-}
-
-// loadCommitFileDiff loads diff for a file in a commit.
-func (p *Plugin) loadCommitFileDiff(hash, path string) tea.Cmd {
-	workDir := p.ctx.WorkDir
-	return func() tea.Msg {
-		rawDiff, err := GetCommitDiff(workDir, hash, path)
-		if err != nil {
-			return ErrorMsg{Err: err}
-		}
-
-		return DiffLoadedMsg{Content: rawDiff, Raw: rawDiff}
-	}
-}
-
 // TickCmd returns a command that triggers a refresh every second.
 func TickCmd() tea.Cmd {
 	return tea.Tick(time.Second, func(t time.Time) tea.Msg {
@@ -2088,220 +1000,13 @@ func (p *Plugin) initCommitTextarea() {
 	p.commitButtonHover = false
 }
 
-// updateCommit handles key events in the commit view.
-func (p *Plugin) updateCommit(msg tea.KeyMsg) (plugin.Plugin, tea.Cmd) {
-	switch msg.String() {
-	case "esc":
-		// Cancel commit, return to status
-		p.viewMode = ViewModeStatus
-		p.commitError = ""
-		return p, nil
 
-	case "ctrl+s":
-		// Execute commit if message not empty
-		return p, p.tryCommit()
-
-	case "tab", "shift+tab":
-		// Toggle focus between textarea and button
-		p.commitButtonFocus = !p.commitButtonFocus
-		if p.commitButtonFocus {
-			p.commitMessage.Blur()
-		} else {
-			p.commitMessage.Focus()
-		}
-		return p, nil
-
-	case "enter":
-		// If button is focused, execute commit
-		if p.commitButtonFocus {
-			return p, p.tryCommit()
-		}
-		// Otherwise let textarea handle it (newline)
-	}
-
-	// Pass other keys to textarea (only if textarea is focused)
-	if !p.commitButtonFocus {
-		var cmd tea.Cmd
-		p.commitMessage, cmd = p.commitMessage.Update(msg)
-		return p, cmd
-	}
-
-	return p, nil
-}
-
-// tryCommit attempts to execute the commit if message is valid.
-func (p *Plugin) tryCommit() tea.Cmd {
-	message := strings.TrimSpace(p.commitMessage.Value())
-	if message == "" {
-		p.commitError = "Commit message cannot be empty"
-		return nil
-	}
-	p.commitInProgress = true
-	return p.doCommit(message)
-}
-
-// updatePushMenu handles key events in the push menu.
-func (p *Plugin) updatePushMenu(msg tea.KeyMsg) (plugin.Plugin, tea.Cmd) {
-	const itemCount = 3 // push, force, upstream
-
-	switch msg.String() {
-	case "tab", "j", "down":
-		p.pushMenuFocus = (p.pushMenuFocus + 1) % itemCount
-		return p, nil
-	case "shift+tab", "k", "up":
-		p.pushMenuFocus = (p.pushMenuFocus - 1 + itemCount) % itemCount
-		return p, nil
-	case "enter":
-		return p.executePushMenuAction(p.pushMenuFocus)
-	case "esc", "q":
-		p.viewMode = p.pushMenuReturnMode
-		p.pushMenuFocus = 0
-		return p, nil
-	case "p":
-		// Regular push (shortcut)
-		return p.executePushMenuAction(0)
-	case "f":
-		// Force push (shortcut)
-		return p.executePushMenuAction(1)
-	case "u":
-		// Push and set upstream (shortcut)
-		return p.executePushMenuAction(2)
-	}
-	return p, nil
-}
-
-// executePushMenuAction executes the push menu action at the given index.
-func (p *Plugin) executePushMenuAction(idx int) (plugin.Plugin, tea.Cmd) {
-	p.viewMode = p.pushMenuReturnMode
-	p.pushInProgress = true
-	p.pushError = ""
-	p.pushSuccess = false
-	p.pushMenuFocus = 0
-
-	// Preserve selected commit hash before push to restore cursor after refresh
-	p.pushPreservedCommitHash = ""
-	if p.cursorOnCommit() {
-		commits := p.activeCommits()
-		commitIdx := p.selectedCommitIndex()
-		if commitIdx >= 0 && commitIdx < len(commits) {
-			p.pushPreservedCommitHash = commits[commitIdx].Hash
-		}
-	}
-
-	switch idx {
-	case 0:
-		return p, p.doPush(false)
-	case 1:
-		return p, p.doPushForce()
-	case 2:
-		return p, p.doPushSetUpstream()
-	}
-	return p, nil
-}
-
-// doCommit executes the git commit asynchronously.
-func (p *Plugin) doCommit(message string) tea.Cmd {
-	workDir := p.ctx.WorkDir
-	return func() tea.Msg {
-		hash, err := ExecuteCommit(workDir, message)
-		if err != nil {
-			return CommitErrorMsg{Err: err}
-		}
-		// Extract first line as subject
-		subject := strings.Split(message, "\n")[0]
-		return CommitSuccessMsg{Hash: hash, Subject: subject}
-	}
-}
-
-// doPush executes a git push asynchronously.
-func (p *Plugin) doPush(force bool) tea.Cmd {
-	workDir := p.ctx.WorkDir
-	return func() tea.Msg {
-		output, err := ExecutePush(workDir, force)
-		if err != nil {
-			return PushErrorMsg{Err: err}
-		}
-		return PushSuccessMsg{Output: output}
-	}
-}
-
-// doPushForce executes a force push with lease.
-func (p *Plugin) doPushForce() tea.Cmd {
-	workDir := p.ctx.WorkDir
-	return func() tea.Msg {
-		output, err := ExecutePushForce(workDir)
-		if err != nil {
-			return PushErrorMsg{Err: err}
-		}
-		return PushSuccessMsg{Output: output}
-	}
-}
-
-// doPushSetUpstream executes a push with upstream tracking.
-func (p *Plugin) doPushSetUpstream() tea.Cmd {
-	workDir := p.ctx.WorkDir
-	return func() tea.Msg {
-		output, err := ExecutePushSetUpstream(workDir)
-		if err != nil {
-			return PushErrorMsg{Err: err}
-		}
-		return PushSuccessMsg{Output: output}
-	}
-}
 
 // clearPushSuccessAfterDelay returns a command that clears the push success indicator after 3 seconds.
 func (p *Plugin) clearPushSuccessAfterDelay() tea.Cmd {
 	return tea.Tick(3*time.Second, func(t time.Time) tea.Msg {
 		return PushSuccessClearMsg{}
 	})
-}
-
-// canPush returns true if there are commits that can be pushed.
-func (p *Plugin) canPush() bool {
-	return p.pushStatus != nil && p.pushStatus.CanPush()
-}
-
-// doStashPush stashes all current changes.
-func (p *Plugin) doStashPush() tea.Cmd {
-	workDir := p.ctx.WorkDir
-	return func() tea.Msg {
-		err := StashPush(workDir)
-		return StashResultMsg{Operation: "push", Err: err}
-	}
-}
-
-// doStashPop pops the latest stash.
-func (p *Plugin) doStashPop() tea.Cmd {
-	workDir := p.ctx.WorkDir
-	return func() tea.Msg {
-		err := StashPop(workDir)
-		return StashResultMsg{Operation: "pop", Ref: "stash@{0}", Err: err}
-	}
-}
-
-
-// doFetch fetches from remote.
-func (p *Plugin) doFetch() tea.Cmd {
-	workDir := p.ctx.WorkDir
-	return func() tea.Msg {
-		output, err := ExecuteFetch(workDir)
-		if err != nil {
-			return FetchErrorMsg{Err: err}
-		}
-		return FetchSuccessMsg{Output: output}
-	}
-}
-
-// doPull pulls from remote.
-func (p *Plugin) doPull() tea.Cmd {
-	workDir := p.ctx.WorkDir
-	return func() tea.Msg {
-		output, err := ExecutePull(workDir)
-		if err != nil {
-			return PullErrorMsg{Err: err}
-		}
-		return PullSuccessMsg{Output: output}
-	}
 }
 
 // clearFetchSuccessAfterDelay returns a command that clears the fetch success indicator after 3 seconds.
@@ -2336,159 +1041,5 @@ type StashPopConfirmMsg struct {
 }
 
 // updateConfirmStashPop handles key events in the confirm stash pop modal.
-// handleConfirmKey handles common key input for confirm dialogs with button pairs.
-// focusPtr points to the button focus state (1=confirm, 2=cancel).
-// onConfirm is called when the user confirms (enter on confirm, y/Y).
-// onCancel is called when the user cancels (enter on cancel, esc/n/N/q).
-func (p *Plugin) handleConfirmKey(msg tea.KeyMsg, focusPtr *int, onConfirm, onCancel func()) (bool, tea.Cmd) {
-	switch msg.String() {
-	case "tab":
-		if *focusPtr == 1 {
-			*focusPtr = 2
-		} else {
-			*focusPtr = 1
-		}
-		return true, nil
-	case "shift+tab":
-		if *focusPtr == 2 {
-			*focusPtr = 1
-		} else {
-			*focusPtr = 2
-		}
-		return true, nil
-	case "enter":
-		if *focusPtr == 2 {
-			onCancel()
-			return true, nil
-		}
-		onConfirm()
-		return true, nil
-	case "y", "Y":
-		onConfirm()
-		return true, nil
-	case "esc", "n", "N", "q":
-		onCancel()
-		return true, nil
-	}
-	return false, nil
-}
 
-func (p *Plugin) updateConfirmStashPop(msg tea.KeyMsg) (plugin.Plugin, tea.Cmd) {
-	var cmd tea.Cmd
-	handled, _ := p.handleConfirmKey(msg, &p.stashPopButtonFocus,
-		func() {
-			// Confirm
-			if p.stashPopItem != nil {
-				cmd = p.doStashPop()
-			}
-			p.viewMode = ViewModeStatus
-			p.stashPopItem = nil
-			p.stashPopButtonFocus = 1
-		},
-		func() {
-			// Cancel
-			p.viewMode = ViewModeStatus
-			p.stashPopItem = nil
-			p.stashPopButtonFocus = 1
-		},
-	)
-	if handled {
-		return p, cmd
-	}
-	return p, nil
-}
 
-// updateConfirmDiscard handles key events in the confirm discard modal.
-func (p *Plugin) updateConfirmDiscard(msg tea.KeyMsg) (plugin.Plugin, tea.Cmd) {
-	var cmd tea.Cmd
-	handled, _ := p.handleConfirmKey(msg, &p.discardButtonFocus,
-		func() {
-			// Confirm
-			if p.discardFile != nil {
-				cmd = p.doDiscard(p.discardFile)
-			}
-			p.viewMode = p.discardReturnMode
-			p.discardFile = nil
-			p.discardButtonFocus = 1
-		},
-		func() {
-			// Cancel
-			p.viewMode = p.discardReturnMode
-			p.discardFile = nil
-			p.discardButtonFocus = 1
-		},
-	)
-	if handled {
-		return p, cmd
-	}
-	return p, nil
-}
-
-// doDiscard executes the git discard operation.
-func (p *Plugin) doDiscard(entry *FileEntry) tea.Cmd {
-	workDir := p.ctx.WorkDir
-	return func() tea.Msg {
-		var err error
-		if entry.Status == StatusUntracked {
-			// Remove untracked file
-			err = DiscardUntracked(workDir, entry.Path)
-		} else if entry.Staged {
-			// Unstage and restore staged file
-			err = DiscardStaged(workDir, entry.Path)
-		} else {
-			// Restore modified file
-			err = DiscardModified(workDir, entry.Path)
-		}
-		if err != nil {
-			return ErrorMsg{Err: err}
-		}
-		return RefreshDoneMsg{}
-	}
-}
-
-// clampDiffHorizScroll clamps diffHorizOff to valid range based on content width.
-func (p *Plugin) clampDiffHorizScroll() {
-	if p.parsedDiff == nil {
-		return
-	}
-	// Calculate content width like the view does
-	panelWidth := (p.width - 3) / 2 // -3 for center separator
-	lineNoWidth := 5
-	contentWidth := panelWidth - lineNoWidth - 2
-
-	clipInfo := GetSideBySideClipInfo(p.parsedDiff, contentWidth, p.diffHorizOff)
-	maxScroll := clipInfo.MaxContentWidth - contentWidth
-	if maxScroll < 0 {
-		maxScroll = 0
-	}
-	if p.diffHorizOff > maxScroll {
-		p.diffHorizOff = maxScroll
-	}
-	if p.diffHorizOff < 0 {
-		p.diffHorizOff = 0
-	}
-}
-
-// clampDiffPaneHorizScroll clamps diffPaneHorizScroll to valid range.
-func (p *Plugin) clampDiffPaneHorizScroll() {
-	if p.diffPaneParsedDiff == nil {
-		return
-	}
-	// Calculate content width for inline diff pane
-	paneWidth := p.width - p.sidebarWidth - 2
-	panelWidth := (paneWidth - 3) / 2
-	lineNoWidth := 5
-	contentWidth := panelWidth - lineNoWidth - 2
-
-	clipInfo := GetSideBySideClipInfo(p.diffPaneParsedDiff, contentWidth, p.diffPaneHorizScroll)
-	maxScroll := clipInfo.MaxContentWidth - contentWidth
-	if maxScroll < 0 {
-		maxScroll = 0
-	}
-	if p.diffPaneHorizScroll > maxScroll {
-		p.diffPaneHorizScroll = maxScroll
-	}
-	if p.diffPaneHorizScroll < 0 {
-		p.diffPaneHorizScroll = 0
-	}
-}
