+++ internal/plugins/worktree/view_modals.go
@@ -3,909 +3,11 @@ package worktree
 import (
 	"fmt"
 	"strings"
-	"time"
 
 	"github.com/charmbracelet/lipgloss"
-	"github.com/charmbracelet/x/ansi"
-	"github.com/marcus/sidecar/internal/plugins/gitstatus"
 	"github.com/marcus/sidecar/internal/styles"
 	"github.com/marcus/sidecar/internal/ui"
 )
-
-var (
-	// Modal styles
-	modalStyle = lipgloss.NewStyle().
-			Border(lipgloss.RoundedBorder()).
-			BorderForeground(lipgloss.Color("62")).
-			Padding(1, 2)
-
-	inputStyle = lipgloss.NewStyle().
-			Border(lipgloss.NormalBorder()).
-			BorderForeground(lipgloss.Color("240")).
-			Padding(0, 1)
-
-	inputFocusedStyle = lipgloss.NewStyle().
-				Border(lipgloss.NormalBorder()).
-				BorderForeground(lipgloss.Color("62")).
-				Padding(0, 1)
-)
-
-const tabStopWidth = 8
-
-// View renders the plugin UI.
-func (p *Plugin) View(width, height int) string {
-	p.width = width
-	p.height = height
-
-	// CRITICAL: Clear hit regions at start of each render
-	p.mouseHandler.Clear()
-
-	switch p.viewMode {
-	case ViewModeCreate:
-		return p.renderCreateModal(width, height)
-	case ViewModeKanban:
-		return p.renderKanbanView(width, height)
-	case ViewModeTaskLink:
-		return p.renderTaskLinkModal(width, height)
-	case ViewModeMerge:
-		return p.renderMergeModal(width, height)
-	case ViewModeAgentChoice:
-		return p.renderAgentChoiceModal(width, height)
-	case ViewModeConfirmDelete:
-		return p.renderConfirmDeleteModal(width, height)
-	case ViewModeCommitForMerge:
-		return p.renderCommitForMergeModal(width, height)
-	case ViewModePromptPicker:
-		return p.renderPromptPickerModal(width, height)
-	default:
-		return p.renderListView(width, height)
-	}
-}
-
-// renderListView renders the main split-pane list view.
-func (p *Plugin) renderListView(width, height int) string {
-	// Pane height for panels (outer dimensions including borders)
-	paneHeight := height
-	if paneHeight < 4 {
-		paneHeight = 4
-	}
-
-	// Inner content height (excluding borders and header lines)
-	innerHeight := paneHeight - 2
-	if innerHeight < 1 {
-		innerHeight = 1
-	}
-
-	// If sidebar is hidden, show only preview pane at full width
-	if !p.sidebarVisible {
-		// Register hit region for full-width preview
-		p.mouseHandler.HitMap.AddRect(regionPreviewPane, 0, 0, width, paneHeight, nil)
-
-		// Register preview tab hit regions (same as sidebar-visible case)
-		// X starts at 2 (1 for border + 1 for panel padding)
-		tabWidths := []int{10, 8, 8} // " Output " + padding, " Diff " + padding, " Task " + padding
-		tabX := 2
-		for i, tabWidth := range tabWidths {
-			p.mouseHandler.HitMap.AddRect(regionPreviewTab, tabX, 1, tabWidth, 1, i)
-			tabX += tabWidth + 1
-		}
-
-		previewContent := p.renderPreviewContent(width-4, innerHeight)
-		return styles.RenderPanel(previewContent, width, paneHeight, true)
-	}
-
-	// RenderPanel handles borders internally, so only subtract divider
-	available := width - dividerWidth
-	sidebarW := (available * p.sidebarWidth) / 100
-	if sidebarW < 25 {
-		sidebarW = 25
-	}
-	if sidebarW > available-40 {
-		sidebarW = available - 40
-	}
-	previewW := available - sidebarW
-	if previewW < 40 {
-		previewW = 40
-	}
-
-	// Determine pane focus state
-	sidebarActive := p.activePane == PaneSidebar
-	previewActive := p.activePane == PanePreview
-
-	// Register hit regions (order matters: last = highest priority)
-	// 1. Pane regions (lowest priority - fallback for scroll)
-	p.mouseHandler.HitMap.AddRect(regionSidebar, 0, 0, sidebarW, paneHeight, nil)
-	p.mouseHandler.HitMap.AddRect(regionPreviewPane, sidebarW+dividerWidth, 0, previewW, paneHeight, nil)
-
-	// 2. Divider region (high priority - for drag)
-	p.mouseHandler.HitMap.AddRect(regionPaneDivider, sidebarW, 0, dividerHitWidth, paneHeight, nil)
-
-	// 3. Preview tab hit regions (highest priority for tabs)
-	// Tabs are rendered at Y=1 (first line inside panel border)
-	// X starts at sidebarW + dividerWidth + 2 (1 for border + 1 for padding)
-	previewPaneX := sidebarW + dividerWidth + 2 // +1 for border, +1 for panel padding
-	// Tab widths: text is " Output " (8), " Diff " (6), " Task " (6)
-	// Plus BarChip Padding(0,1) adds 2 chars = 10, 8, 8 visual width
-	tabWidths := []int{10, 8, 8}
-	tabX := previewPaneX
-	for i, tabWidth := range tabWidths {
-		p.mouseHandler.HitMap.AddRect(regionPreviewTab, tabX, 1, tabWidth, 1, i)
-		tabX += tabWidth + 1 // +1 for spacing between tabs
-	}
-
-	// Render content for each pane (subtract 4 for border + padding: 2 border + 2 padding)
-	sidebarContent := p.renderSidebarContent(sidebarW-4, innerHeight)
-	previewContent := p.renderPreviewContent(previewW-4, innerHeight)
-
-	// Apply gradient border styles
-	leftPane := styles.RenderPanel(sidebarContent, sidebarW, paneHeight, sidebarActive)
-	rightPane := styles.RenderPanel(previewContent, previewW, paneHeight, previewActive)
-
-	// Render visible divider between panes
-	divider := p.renderDivider(paneHeight)
-
-	// Join horizontally
-	return lipgloss.JoinHorizontal(lipgloss.Top, leftPane, divider, rightPane)
-}
-
-// renderSidebarContent renders the worktree list sidebar content (no borders).
-func (p *Plugin) renderSidebarContent(width, height int) string {
-	var lines []string
-
-	// Header
-	header := styles.Title.Render("Worktrees")
-	lines = append(lines, header)
-	lines = append(lines, "") // Empty line after header
-
-	// Track Y position for hit regions (add 3 for border + header + empty line)
-	currentY := 3
-
-	// Calculate visible items (each item is 2 lines)
-	contentHeight := height - 2 // header + empty line
-	itemHeight := 2             // Each worktree item takes 2 lines
-	p.visibleCount = contentHeight / itemHeight
-
-	// Render worktree items
-	if len(p.worktrees) == 0 {
-		// Calculate vertical centering for empty state
-		emptyStateHeight := 2 // "No worktrees" + "Press 'n'..."
-		emptyStartY := (contentHeight - emptyStateHeight) / 2
-		if emptyStartY < 0 {
-			emptyStartY = 0
-		}
-
-		// Add padding lines before empty state message
-		for i := 0; i < emptyStartY; i++ {
-			lines = append(lines, "")
-		}
-
-		// Center the text horizontally
-		msg1 := "No worktrees"
-		msg2 := "Press 'n' to create one"
-		pad1 := (width - len(msg1)) / 2
-		pad2 := (width - len(msg2)) / 2
-		if pad1 < 0 {
-			pad1 = 0
-		}
-		if pad2 < 0 {
-			pad2 = 0
-		}
-
-		lines = append(lines, styles.Muted.Render(strings.Repeat(" ", pad1)+msg1))
-		lines = append(lines, styles.Muted.Render(strings.Repeat(" ", pad2)+msg2))
-	} else {
-		for i := p.scrollOffset; i < len(p.worktrees) && i < p.scrollOffset+p.visibleCount; i++ {
-			wt := p.worktrees[i]
-			line := p.renderWorktreeItem(wt, i == p.selectedIdx, width)
-
-			// Register hit region with ABSOLUTE index
-			p.mouseHandler.HitMap.AddRect(regionWorktreeItem, 0, currentY, width, itemHeight, i)
-
-			lines = append(lines, line)
-			currentY += itemHeight
-		}
-	}
-
-	return strings.Join(lines, "\n")
-}
-
-// renderWorktreeItem renders a single worktree list item.
-func (p *Plugin) renderWorktreeItem(wt *Worktree, selected bool, width int) string {
-	// Keep selection visible even when preview pane is active (dimmed style)
-	isSelected := selected
-	isActiveFocus := selected && p.activePane == PaneSidebar
-
-	// Status indicator
-	statusIcon := wt.Status.Icon()
-
-	// Check for conflicts
-	hasConflict := p.hasConflict(wt.Name, p.conflicts)
-	conflictIcon := ""
-	if hasConflict {
-		conflictIcon = " ⚠"
-	}
-
-	// Check for PR
-	hasPR := wt.PRURL != ""
-	prIcon := ""
-	if hasPR {
-		prIcon = " PR"
-	}
-
-	// Name and time
-	name := wt.Name
-	timeStr := formatRelativeTime(wt.UpdatedAt)
-
-	// Stats if available
-	statsStr := ""
-	if wt.Stats != nil && (wt.Stats.Additions > 0 || wt.Stats.Deletions > 0) {
-		statsStr = fmt.Sprintf("+%d -%d", wt.Stats.Additions, wt.Stats.Deletions)
-	}
-
-	// Build second line parts (plain text)
-	var parts []string
-	if wt.Agent != nil {
-		parts = append(parts, string(wt.Agent.Type))
-	} else if wt.ChosenAgentType != "" && wt.ChosenAgentType != AgentNone {
-		parts = append(parts, string(wt.ChosenAgentType))
-	} else {
-		parts = append(parts, "—")
-	}
-	if wt.TaskID != "" {
-		parts = append(parts, wt.TaskID)
-	}
-	if statsStr != "" {
-		parts = append(parts, statsStr)
-	}
-	if hasConflict {
-		conflictFiles := p.getConflictingFiles(wt.Name, p.conflicts)
-		if len(conflictFiles) > 0 {
-			parts = append(parts, fmt.Sprintf("⚠ %d conflicts", len(conflictFiles)))
-		}
-	}
-
-	// When selected, use plain text to ensure consistent background
-	if isSelected {
-		// Build plain text lines
-		line1 := fmt.Sprintf(" %s %s%s%s", statusIcon, name, prIcon, conflictIcon)
-		line1Width := lipgloss.Width(line1)
-		timeWidth := lipgloss.Width(timeStr)
-		if line1Width < width-timeWidth-2 {
-			line1 = line1 + strings.Repeat(" ", width-line1Width-timeWidth-1) + timeStr
-		}
-		line2 := "   " + strings.Join(parts, "  ")
-		// Pad line2 to full width for consistent background
-		line2Width := lipgloss.Width(line2)
-		if line2Width < width {
-			line2 = line2 + strings.Repeat(" ", width-line2Width)
-		}
-		content := line1 + "\n" + line2
-
-		// Use bright style when sidebar is focused, dimmed when preview is focused
-		if isActiveFocus {
-			return styles.ListItemSelected.Width(width).Render(content)
-		}
-		// Dimmed selection style (when preview pane is active)
-		dimmedSelectedStyle := lipgloss.NewStyle().
-			Background(lipgloss.Color("237")). // Darker background
-			Foreground(lipgloss.Color("252")). // Slightly dimmed text
-			Width(width)
-		return dimmedSelectedStyle.Render(content)
-	}
-
-	// Not selected - use colored styles for visual interest
-	var statusStyle lipgloss.Style
-	switch wt.Status {
-	case StatusActive:
-		statusStyle = styles.StatusCompleted // Green
-	case StatusWaiting:
-		statusStyle = styles.StatusModified // Yellow/orange (warning)
-	case StatusDone:
-		statusStyle = styles.StatusCompleted // Green
-	case StatusError:
-		statusStyle = styles.StatusDeleted // Red
-	default:
-		statusStyle = styles.Muted // Gray for paused
-	}
-	icon := statusStyle.Render(statusIcon)
-
-	// Apply conflict style
-	styledConflictIcon := ""
-	if hasConflict {
-		styledConflictIcon = styles.StatusModified.Render(" ⚠")
-	}
-
-	// Apply PR style
-	styledPRIcon := ""
-	if hasPR {
-		styledPRIcon = lipgloss.NewStyle().Foreground(lipgloss.Color("33")).Render(" PR") // blue
-	}
-
-	// For non-selected, style parts individually
-	var styledParts []string
-	if wt.Agent != nil {
-		styledParts = append(styledParts, string(wt.Agent.Type))
-	} else if wt.ChosenAgentType != "" && wt.ChosenAgentType != AgentNone {
-		styledParts = append(styledParts, dimText(string(wt.ChosenAgentType)))
-	} else {
-		styledParts = append(styledParts, "—")
-	}
-	if wt.TaskID != "" {
-		styledParts = append(styledParts, wt.TaskID)
-	}
-	if statsStr != "" {
-		styledParts = append(styledParts, statsStr)
-	}
-	if hasConflict {
-		conflictFiles := p.getConflictingFiles(wt.Name, p.conflicts)
-		if len(conflictFiles) > 0 {
-			styledParts = append(styledParts, styles.StatusModified.Render(fmt.Sprintf("⚠ %d conflicts", len(conflictFiles))))
-		}
-	}
-
-	// Build lines with styled elements
-	line1 := fmt.Sprintf(" %s %s%s%s", icon, name, styledPRIcon, styledConflictIcon)
-	line1Width := ansi.StringWidth(line1)
-	timeWidth := ansi.StringWidth(timeStr)
-	if line1Width < width-timeWidth-2 {
-		line1 = line1 + strings.Repeat(" ", width-line1Width-timeWidth-1) + timeStr
-	}
-	line2 := "   " + strings.Join(styledParts, "  ")
-
-	content := line1 + "\n" + line2
-	return styles.ListItemNormal.Width(width).Render(content)
-}
-
-// renderPreviewContent renders the preview pane content (no borders).
-func (p *Plugin) renderPreviewContent(width, height int) string {
-	var lines []string
-
-	// Hide tabs when no worktree is selected - show welcome guide instead
-	wt := p.selectedWorktree()
-	if wt == nil {
-		return truncateAllLines(p.renderWelcomeGuide(width, height), width)
-	}
-
-	// Tab header
-	tabs := p.renderTabs(width)
-	lines = append(lines, tabs)
-	lines = append(lines, "") // Empty line after header
-
-	contentHeight := height - 2 // header + empty line
-
-	// Render content based on active tab
-	var content string
-	switch p.previewTab {
-	case PreviewTabOutput:
-		content = p.renderOutputContent(width, contentHeight)
-	case PreviewTabDiff:
-		content = p.renderDiffContent(width, contentHeight)
-	case PreviewTabTask:
-		content = p.renderTaskContent(width, contentHeight)
-	}
-
-	lines = append(lines, content)
-
-	// Final safety: ensure ALL lines are truncated to width
-	// This catches any content that wasn't properly truncated
-	result := strings.Join(lines, "\n")
-	return truncateAllLines(result, width)
-}
-
-// renderWelcomeGuide renders a helpful guide when no worktree is selected.
-func (p *Plugin) renderWelcomeGuide(width, height int) string {
-	var lines []string
-
-	// Section Style
-	sectionStyle := lipgloss.NewStyle().Bold(true).Foreground(lipgloss.Color("62"))
-
-	// Git Worktree Explanation
-	lines = append(lines, sectionStyle.Render("Git Worktrees: A Better Workflow"))
-	lines = append(lines, dimText("  • Parallel Development: Work on multiple branches simultaneously"))
-	lines = append(lines, dimText("    in separate directories."))
-	lines = append(lines, dimText("  • No Context Switching: Keep your editor/server running while"))
-	lines = append(lines, dimText("    reviewing a PR or fixing a bug."))
-	lines = append(lines, dimText("  • Isolated Environments: Each worktree has its own clean state,"))
-	lines = append(lines, dimText("    unaffected by other changes."))
-	lines = append(lines, "")
-	lines = append(lines, strings.Repeat("─", min(width-4, 60)))
-	lines = append(lines, "")
-
-	// Title
-	title := lipgloss.NewStyle().Bold(true).Render("tmux Quick Reference")
-	lines = append(lines, title)
-	lines = append(lines, "")
-
-	// Section: Attaching to agent sessions
-	lines = append(lines, sectionStyle.Render("Agent Sessions"))
-	lines = append(lines, dimText("  Enter      Attach to selected worktree session"))
-	lines = append(lines, dimText("  Ctrl-b d   Detach from session (return here)"))
-	lines = append(lines, "")
-
-	// Section: Navigation inside tmux
-	lines = append(lines, sectionStyle.Render("Scrolling (in attached session)"))
-	lines = append(lines, dimText("  Ctrl-b [        Enter scroll mode"))
-	lines = append(lines, dimText("  PgUp/PgDn       Scroll page (fn+↑/↓ on Mac)"))
-	lines = append(lines, dimText("  ↑/↓             Scroll line by line"))
-	lines = append(lines, dimText("  q               Exit scroll mode"))
-	lines = append(lines, "")
-
-	// Section: Interacting with editors
-	lines = append(lines, sectionStyle.Render("Editor Navigation"))
-	lines = append(lines, dimText("  When agent opens vim/nano:"))
-	lines = append(lines, dimText("    :q!      Quit vim without saving"))
-	lines = append(lines, dimText("    :wq      Save and quit vim"))
-	lines = append(lines, dimText("    Ctrl-x   Exit nano"))
-	lines = append(lines, "")
-
-	// Section: Common tasks
-	lines = append(lines, sectionStyle.Render("Tips"))
-	lines = append(lines, dimText("  • Create a worktree with 'n' to start"))
-	lines = append(lines, dimText("  • Agent output streams in the Output tab"))
-	lines = append(lines, dimText("  • Attach to interact with the agent directly"))
-	lines = append(lines, "")
-	lines = append(lines, dimText("Customize tmux: ~/.tmux.conf (man tmux for options)"))
-
-	return strings.Join(lines, "\n")
-}
-
-// truncateAllLines ensures every line in the content is truncated to maxWidth.
-func truncateAllLines(content string, maxWidth int) string {
-	lines := strings.Split(content, "\n")
-	for i, line := range lines {
-		line = expandTabs(line, tabStopWidth)
-		if lipgloss.Width(line) > maxWidth {
-			line = ansi.Truncate(line, maxWidth, "")
-		}
-		lines[i] = line
-	}
-	return strings.Join(lines, "\n")
-}
-
-// expandTabs replaces tabs with spaces, preserving ANSI sequences and column widths.
-func expandTabs(line string, tabWidth int) string {
-	if tabWidth <= 0 || !strings.Contains(line, "\t") {
-		return line
-	}
-
-	var sb strings.Builder
-	sb.Grow(len(line))
-
-	state := ansi.NormalState
-	column := 0
-	for len(line) > 0 {
-		seq, width, n, newState := ansi.GraphemeWidth.DecodeSequenceInString(line, state, nil)
-		if n <= 0 {
-			sb.WriteString(line)
-			break
-		}
-		if seq == "\t" && width == 0 {
-			spaces := tabWidth - (column % tabWidth)
-			if spaces == 0 {
-				spaces = tabWidth
-			}
-			sb.WriteString(strings.Repeat(" ", spaces))
-			column += spaces
-		} else {
-			sb.WriteString(seq)
-			column += width
-		}
-		state = newState
-		line = line[n:]
-	}
-
-	return sb.String()
-}
-
-// renderTabs renders the preview pane tab header.
-func (p *Plugin) renderTabs(width int) string {
-	tabs := []string{"Output", "Diff", "Task"}
-	var rendered []string
-
-	for i, tab := range tabs {
-		if PreviewTab(i) == p.previewTab {
-			rendered = append(rendered, styles.BarChipActive.Render(" "+tab+" "))
-		} else {
-			rendered = append(rendered, styles.BarChip.Render(" "+tab+" "))
-		}
-	}
-
-	return strings.Join(rendered, " ")
-}
-
-// renderOutputContent renders agent output.
-func (p *Plugin) renderOutputContent(width, height int) string {
-	wt := p.selectedWorktree()
-	if wt == nil {
-		return dimText("No worktree selected")
-	}
-
-	if wt.Agent == nil {
-		return dimText("No agent running\nPress 's' to start an agent")
-	}
-
-	// Hint for tmux detach
-	hint := dimText("enter to attach • Ctrl-b d to detach")
-	height-- // Reserve line for hint
-
-	if wt.Agent.OutputBuf == nil {
-		return hint + "\n" + dimText("No output yet")
-	}
-
-	lines := wt.Agent.OutputBuf.Lines()
-	if len(lines) == 0 {
-		return hint + "\n" + dimText("No output yet")
-	}
-
-	var start, end int
-	if p.autoScrollOutput {
-		// Auto-scroll: show newest content (last height lines)
-		start = len(lines) - height
-		if start < 0 {
-			start = 0
-		}
-		end = len(lines)
-	} else {
-		// Manual scroll: previewOffset is lines from bottom
-		// offset=0 means bottom, offset=N means N lines up from bottom
-		start = len(lines) - height - p.previewOffset
-		if start < 0 {
-			start = 0
-		}
-		end = start + height
-		if end > len(lines) {
-			end = len(lines)
-		}
-	}
-
-	// Apply horizontal offset and truncate each line
-	var displayLines []string
-	for _, line := range lines[start:end] {
-		displayLine := expandTabs(line, tabStopWidth)
-		// Apply horizontal offset using ANSI-aware truncation
-		if p.previewHorizOffset > 0 {
-			displayLine = ansi.TruncateLeft(displayLine, p.previewHorizOffset, "")
-		}
-		// Truncate to width if needed
-		if lipgloss.Width(displayLine) > width {
-			displayLine = ansi.Truncate(displayLine, width, "")
-		}
-		displayLines = append(displayLines, displayLine)
-	}
-
-	return hint + "\n" + strings.Join(displayLines, "\n")
-}
-
-// renderCommitStatusHeader renders the commit status header for diff view.
-func (p *Plugin) renderCommitStatusHeader(width int) string {
-	if len(p.commitStatusList) == 0 {
-		return ""
-	}
-
-	// Box style for header
-	headerStyle := lipgloss.NewStyle().
-		Border(lipgloss.RoundedBorder()).
-		BorderForeground(lipgloss.Color("62")).
-		Padding(0, 1).
-		Width(width - 2)
-
-	titleStyle := lipgloss.NewStyle().Bold(true).Foreground(lipgloss.Color("62"))
-	hashStyle := lipgloss.NewStyle().Foreground(lipgloss.Color("214"))
-	pushedStyle := lipgloss.NewStyle().Foreground(lipgloss.Color("42"))
-	localStyle := lipgloss.NewStyle().Foreground(lipgloss.Color("240"))
-
-	var sb strings.Builder
-	sb.WriteString(titleStyle.Render(fmt.Sprintf("Commits (%d)", len(p.commitStatusList))))
-	sb.WriteString("\n")
-
-	// Show up to 5 commits
-	maxCommits := 5
-	displayCount := len(p.commitStatusList)
-	if displayCount > maxCommits {
-		displayCount = maxCommits
-	}
-
-	for i := 0; i < displayCount; i++ {
-		commit := p.commitStatusList[i]
-
-		// Status icon
-		var statusIcon string
-		if commit.Pushed {
-			statusIcon = pushedStyle.Render("↑")
-		} else {
-			statusIcon = localStyle.Render("○")
-		}
-
-		// Truncate subject to fit
-		subject := commit.Subject
-		maxSubjectLen := width - 15 // hash(7) + icon(2) + spaces(6)
-		if maxSubjectLen < 10 {
-			maxSubjectLen = 10
-		}
-		if len(subject) > maxSubjectLen {
-			subject = subject[:maxSubjectLen-3] + "..."
-		}
-
-		line := fmt.Sprintf("%s %s %s", statusIcon, hashStyle.Render(commit.Hash), subject)
-		sb.WriteString(line)
-		if i < displayCount-1 {
-			sb.WriteString("\n")
-		}
-	}
-
-	if len(p.commitStatusList) > maxCommits {
-		sb.WriteString("\n")
-		sb.WriteString(dimText(fmt.Sprintf("  ... and %d more", len(p.commitStatusList)-maxCommits)))
-	}
-
-	return headerStyle.Render(sb.String())
-}
-
-// renderDiffContent renders git diff using the shared diff renderer.
-func (p *Plugin) renderDiffContent(width, height int) string {
-	wt := p.selectedWorktree()
-	if wt == nil {
-		return dimText("No worktree selected")
-	}
-
-	// Render commit status header if it belongs to current worktree
-	header := ""
-	if p.commitStatusWorktree == wt.Name {
-		header = p.renderCommitStatusHeader(width)
-	}
-
-	headerHeight := 0
-	if header != "" {
-		headerHeight = lipgloss.Height(header) + 1 // +1 for blank line
-	}
-
-	if p.diffRaw == "" {
-		if header != "" {
-			return header + "\n" + dimText("No uncommitted changes")
-		}
-		return dimText("No changes")
-	}
-
-	// Adjust available height for diff content
-	contentHeight := height - headerHeight
-	if contentHeight < 5 {
-		contentHeight = 5
-	}
-
-	// Parse the raw diff into structured format
-	parsed, err := gitstatus.ParseUnifiedDiff(p.diffRaw)
-	if err != nil || parsed == nil {
-		// Fallback to basic rendering
-		diffContent := p.renderDiffContentBasicWithHeight(width, contentHeight)
-		if header != "" {
-			return header + "\n" + diffContent
-		}
-		return diffContent
-	}
-
-	// Create syntax highlighter if we have file info
-	var highlighter *gitstatus.SyntaxHighlighter
-	if parsed.NewFile != "" {
-		highlighter = gitstatus.NewSyntaxHighlighter(parsed.NewFile)
-	}
-
-	// Render based on view mode
-	var diffContent string
-	if p.diffViewMode == DiffViewSideBySide {
-		diffContent = gitstatus.RenderSideBySide(parsed, width, p.previewOffset, contentHeight, p.previewHorizOffset, highlighter)
-	} else {
-		diffContent = gitstatus.RenderLineDiff(parsed, width, p.previewOffset, contentHeight, p.previewHorizOffset, highlighter)
-	}
-
-	if header != "" {
-		return header + "\n" + diffContent
-	}
-	return diffContent
-}
-
-// renderDiffContentBasic renders git diff with basic highlighting (fallback).
-func (p *Plugin) renderDiffContentBasic(width, height int) string {
-	return p.renderDiffContentBasicWithHeight(width, height)
-}
-
-// renderDiffContentBasicWithHeight renders git diff with basic highlighting with explicit height.
-func (p *Plugin) renderDiffContentBasicWithHeight(width, height int) string {
-	lines := splitLines(p.diffContent)
-
-	// Apply scroll offset
-	start := p.previewOffset
-	if start >= len(lines) {
-		start = len(lines) - 1
-	}
-	if start < 0 {
-		start = 0
-	}
-	end := start + height
-	if end > len(lines) {
-		end = len(lines)
-	}
-
-	// Diff highlighting with horizontal scroll support
-	var rendered []string
-	for _, line := range lines[start:end] {
-		line = expandTabs(line, tabStopWidth)
-		var styledLine string
-		switch {
-		case strings.HasPrefix(line, "+++") || strings.HasPrefix(line, "---"):
-			styledLine = styles.DiffHeader.Render(line)
-		case strings.HasPrefix(line, "@@"):
-			styledLine = lipgloss.NewStyle().Foreground(styles.Info).Render(line)
-		case strings.HasPrefix(line, "+"):
-			styledLine = styles.DiffAdd.Render(line)
-		case strings.HasPrefix(line, "-"):
-			styledLine = styles.DiffRemove.Render(line)
-		default:
-			styledLine = line
-		}
-
-		if p.previewHorizOffset > 0 {
-			styledLine = ansi.TruncateLeft(styledLine, p.previewHorizOffset, "")
-		}
-		if lipgloss.Width(styledLine) > width {
-			styledLine = ansi.Truncate(styledLine, width, "")
-		}
-		rendered = append(rendered, styledLine)
-	}
-
-	return strings.Join(rendered, "\n")
-}
-
-// colorDiffLine applies basic diff coloring using theme styles.
-func colorDiffLine(line string, width int) string {
-	line = expandTabs(line, tabStopWidth)
-	if len(line) == 0 {
-		return line
-	}
-
-	// Truncate if needed
-	if lipgloss.Width(line) > width {
-		line = ansi.Truncate(line, width, "")
-	}
-
-	switch {
-	case strings.HasPrefix(line, "+++") || strings.HasPrefix(line, "---"):
-		return styles.DiffHeader.Render(line)
-	case strings.HasPrefix(line, "@@"):
-		return lipgloss.NewStyle().Foreground(styles.Info).Render(line)
-	case strings.HasPrefix(line, "+"):
-		return styles.DiffAdd.Render(line)
-	case strings.HasPrefix(line, "-"):
-		return styles.DiffRemove.Render(line)
-	default:
-		return line
-	}
-}
-
-// renderTaskContent renders linked task info.
-func (p *Plugin) renderTaskContent(width, height int) string {
-	wt := p.selectedWorktree()
-	if wt == nil {
-		return dimText("No worktree selected")
-	}
-
-	if wt.TaskID == "" {
-		return dimText("No linked task\nPress 't' to link a task")
-	}
-
-	// Check if we have cached details for this task
-	if p.cachedTask == nil || p.cachedTaskID != wt.TaskID {
-		return dimText(fmt.Sprintf("Loading task %s...", wt.TaskID))
-	}
-
-	task := p.cachedTask
-	var lines []string
-
-	// Mode indicator
-	modeHint := dimText("[m] raw")
-	if p.taskMarkdownMode {
-		modeHint = dimText("[m] rendered")
-	}
-
-	// Header
-	lines = append(lines, lipgloss.NewStyle().Bold(true).Render(fmt.Sprintf("Task: %s", task.ID))+"  "+modeHint)
-
-	// Status and priority
-	statusLine := fmt.Sprintf("Status: %s", task.Status)
-	if task.Priority != "" {
-		statusLine += fmt.Sprintf("  Priority: %s", task.Priority)
-	}
-	if task.Type != "" {
-		statusLine += fmt.Sprintf("  Type: %s", task.Type)
-	}
-	lines = append(lines, statusLine)
-	lines = append(lines, strings.Repeat("─", min(width-4, 60)))
-	lines = append(lines, "")
-
-	// Title
-	lines = append(lines, lipgloss.NewStyle().Bold(true).Render(task.Title))
-	lines = append(lines, "")
-
-	// Markdown rendering for description and acceptance
-	if p.taskMarkdownMode && p.markdownRenderer != nil {
-		// Build markdown content
-		var mdContent strings.Builder
-		if task.Description != "" {
-			mdContent.WriteString(task.Description)
-			mdContent.WriteString("\n\n")
-		}
-		if task.Acceptance != "" {
-			mdContent.WriteString("## Acceptance Criteria\n\n")
-			mdContent.WriteString(task.Acceptance)
-		}
-
-		// Check if we need to re-render (width changed or cache empty)
-		if p.taskMarkdownWidth != width || len(p.taskMarkdownRendered) == 0 {
-			p.taskMarkdownRendered = p.markdownRenderer.RenderContent(mdContent.String(), width-4)
-			p.taskMarkdownWidth = width
-		}
-
-		// Append rendered lines
-		lines = append(lines, p.taskMarkdownRendered...)
-	} else {
-		// Plain text fallback
-		if task.Description != "" {
-			wrapped := wrapText(task.Description, width-4)
-			lines = append(lines, wrapped)
-			lines = append(lines, "")
-		}
-
-		if task.Acceptance != "" {
-			lines = append(lines, lipgloss.NewStyle().Bold(true).Render("Acceptance Criteria:"))
-			wrapped := wrapText(task.Acceptance, width-4)
-			lines = append(lines, wrapped)
-			lines = append(lines, "")
-		}
-	}
-
-	// Timestamps (dimmed)
-	lines = append(lines, "")
-	if task.CreatedAt != "" {
-		lines = append(lines, dimText(fmt.Sprintf("Created: %s", task.CreatedAt)))
-	}
-	if task.UpdatedAt != "" {
-		lines = append(lines, dimText(fmt.Sprintf("Updated: %s", task.UpdatedAt)))
-	}
-
-	return strings.Join(lines, "\n")
-}
-
-// wrapText wraps text to the specified width.
-func wrapText(text string, width int) string {
-	if width <= 0 {
-		return text
-	}
-
-	var lines []string
-	for _, para := range strings.Split(text, "\n") {
-		if len(para) <= width {
-			lines = append(lines, para)
-			continue
-		}
-
-		// Simple word wrapping
-		words := strings.Fields(para)
-		var currentLine string
-		for _, word := range words {
-			if currentLine == "" {
-				currentLine = word
-			} else if len(currentLine)+1+len(word) <= width {
-				currentLine += " " + word
-			} else {
-				lines = append(lines, currentLine)
-				currentLine = word
-			}
-		}
-		if currentLine != "" {
-			lines = append(lines, currentLine)
-		}
-	}
-	return strings.Join(lines, "\n")
-}
 
 // renderCreateModal renders the new worktree modal with dimmed background.
 func (p *Plugin) renderCreateModal(width, height int) string {
@@ -1732,11 +834,6 @@ func (p *Plugin) renderAgentChoiceModal(width, height int) string {
 	return ui.OverlayModal(background, modal, width, height)
 }
 
-// dimText renders dim placeholder text using theme style.
-func dimText(s string) string {
-	return styles.Muted.Render(s)
-}
-
 // renderMergeModal renders the merge workflow modal with dimmed background.
 func (p *Plugin) renderMergeModal(width, height int) string {
 	// Render the background (list view)
@@ -2149,250 +1246,4 @@ func (p *Plugin) renderCommitForMergeModal(width, height int) string {
 
 	// Use OverlayModal for dimmed background effect
 	return ui.OverlayModal(background, modal, width, height)
-}
-
-// formatRelativeTime formats a time as relative (e.g., "3m", "2h").
-func formatRelativeTime(t time.Time) string {
-	if t.IsZero() {
-		return ""
-	}
-
-	d := time.Since(t)
-	switch {
-	case d < time.Minute:
-		return "now"
-	case d < time.Hour:
-		return fmt.Sprintf("%dm", int(d.Minutes()))
-	case d < 24*time.Hour:
-		return fmt.Sprintf("%dh", int(d.Hours()))
-	default:
-		return fmt.Sprintf("%dd", int(d.Hours()/24))
-	}
-}
-
-// renderDivider renders the vertical divider between panes.
-func (p *Plugin) renderDivider(height int) string {
-	// Use a subtle vertical bar as the divider with theme color
-	// MarginTop(1) shifts it down to align with pane content (below top border)
-	dividerStyle := lipgloss.NewStyle().
-		Foreground(styles.BorderNormal).
-		MarginTop(1)
-
-	var sb strings.Builder
-	for i := 0; i < height; i++ {
-		sb.WriteString("│")
-		if i < height-1 {
-			sb.WriteString("\n")
-		}
-	}
-	return dividerStyle.Render(sb.String())
-}
-
-// renderKanbanView renders the kanban board view.
-func (p *Plugin) renderKanbanView(width, height int) string {
-	// Check minimum width - auto-collapse to list view if too narrow
-	minKanbanWidth := 80
-	if width < minKanbanWidth {
-		// Fall back to list view when too narrow
-		return p.renderListView(width, height)
-	}
-
-	// Use styled separator characters for theme consistency
-	borderStyle := lipgloss.NewStyle().Foreground(styles.BorderNormal)
-	horizSep := borderStyle.Render("─")
-	vertSep := borderStyle.Render("│")
-
-	var lines []string
-
-	// Header with view mode toggle (account for panel border width)
-	innerWidth := width - 4 // Account for panel borders
-	header := styles.Title.Render("Worktrees")
-	listTab := "List"
-	kanbanTab := "[Kanban]"
-	viewToggle := styles.Muted.Render(listTab + "|" + kanbanTab)
-	headerLine := header + strings.Repeat(" ", max(1, innerWidth-len("Worktrees")-len(listTab)-len(kanbanTab)-1)) + viewToggle
-	lines = append(lines, headerLine)
-	lines = append(lines, strings.Repeat(horizSep, innerWidth))
-
-	// Register view toggle hit regions (inside panel border at Y=1)
-	// Position: right-aligned in header line
-	toggleTotalWidth := len(listTab) + 1 + len(kanbanTab) // "List|[Kanban]"
-	toggleX := width - 2 - toggleTotalWidth               // -2 for panel border
-	p.mouseHandler.HitMap.AddRect(regionViewToggle, toggleX, 1, len(listTab), 1, 0)
-	p.mouseHandler.HitMap.AddRect(regionViewToggle, toggleX+len(listTab)+1, 1, len(kanbanTab), 1, 1)
-
-	// Group worktrees by status
-	columns := p.getKanbanColumns()
-
-	// Column headers and colors
-	columnTitles := map[WorktreeStatus]string{
-		StatusActive:  "● Active",
-		StatusWaiting: "💬 Waiting",
-		StatusDone:    "✓ Ready",
-		StatusPaused:  "⏸ Paused",
-	}
-	columnColors := map[WorktreeStatus]lipgloss.Color{
-		StatusActive:  styles.StatusCompleted.GetForeground().(lipgloss.Color), // Green
-		StatusWaiting: styles.StatusModified.GetForeground().(lipgloss.Color),  // Yellow
-		StatusDone:    lipgloss.Color("81"),                                    // Cyan
-		StatusPaused:  lipgloss.Color("245"),                                   // Gray
-	}
-
-	// Calculate column widths (account for panel borders)
-	numCols := len(kanbanColumnOrder)
-	colWidth := (innerWidth - numCols - 1) / numCols // -1 for separators
-	if colWidth < 18 {
-		colWidth = 18
-	}
-
-	// Render column headers with colors and register hit regions
-	var colHeaders []string
-	colX := 2 // Start after panel border
-	for colIdx, status := range kanbanColumnOrder {
-		items := columns[status]
-		title := fmt.Sprintf("%s (%d)", columnTitles[status], len(items))
-		headerStyle := lipgloss.NewStyle().Bold(true).Foreground(columnColors[status]).Width(colWidth)
-		// Highlight selected column header
-		if colIdx == p.kanbanCol {
-			headerStyle = headerStyle.Underline(true)
-		}
-		colHeaders = append(colHeaders, headerStyle.Render(title))
-
-		// Register column header hit region (Y=3, after header line, separator line)
-		p.mouseHandler.HitMap.AddRect(regionKanbanColumn, colX, 3, colWidth, 1, colIdx)
-		colX += colWidth + 1 // +1 for separator
-	}
-	lines = append(lines, strings.Join(colHeaders, vertSep))
-	lines = append(lines, strings.Repeat(horizSep, innerWidth))
-
-	// Card dimensions: 4 lines per card (name, agent, task, stats)
-	cardHeight := 4
-	// Calculate content height (account for panel border + header + separators)
-	contentHeight := height - 6 // panel borders (2) + header + 2 separators + column headers
-	if contentHeight < cardHeight {
-		contentHeight = cardHeight
-	}
-	maxCards := contentHeight / cardHeight
-
-	// Find the maximum number of cards in any column (for row rendering)
-	maxInColumn := 0
-	for _, status := range kanbanColumnOrder {
-		if len(columns[status]) > maxInColumn {
-			maxInColumn = len(columns[status])
-		}
-	}
-	if maxInColumn > maxCards {
-		maxInColumn = maxCards
-	}
-
-	// Render cards row by row and register card hit regions
-	// Cards start at Y=5 (panel border(1) + header(1) + sep(1) + col headers(1) + sep(1))
-	cardStartY := 5
-	for cardIdx := 0; cardIdx < maxInColumn; cardIdx++ {
-		// Register hit regions for this row of cards (once per card, not per line)
-		cardColX := 2 // Start after panel border
-		for colIdx, status := range kanbanColumnOrder {
-			items := columns[status]
-			if cardIdx < len(items) {
-				cardY := cardStartY + (cardIdx * cardHeight)
-				p.mouseHandler.HitMap.AddRect(regionKanbanCard, cardColX, cardY, colWidth-1, cardHeight, kanbanCardData{col: colIdx, row: cardIdx})
-			}
-			cardColX += colWidth + 1 // +1 for separator
-		}
-
-		// Each card has 4 lines
-		for lineIdx := 0; lineIdx < cardHeight; lineIdx++ {
-			var rowCells []string
-			for colIdx, status := range kanbanColumnOrder {
-				items := columns[status]
-				var cellContent string
-
-				if cardIdx < len(items) {
-					wt := items[cardIdx]
-					isSelected := colIdx == p.kanbanCol && cardIdx == p.kanbanRow
-					cellContent = p.renderKanbanCardLine(wt, lineIdx, colWidth-1, isSelected)
-				} else {
-					cellContent = strings.Repeat(" ", colWidth-1)
-				}
-
-				rowCells = append(rowCells, cellContent)
-			}
-			lines = append(lines, strings.Join(rowCells, vertSep))
-		}
-	}
-
-	// Fill remaining height with empty space
-	renderedRows := maxInColumn * cardHeight
-	for i := renderedRows; i < contentHeight; i++ {
-		var emptyCells []string
-		for range kanbanColumnOrder {
-			emptyCells = append(emptyCells, strings.Repeat(" ", colWidth-1))
-		}
-		lines = append(lines, strings.Join(emptyCells, vertSep))
-	}
-
-	// Build content for panel
-	content := strings.Join(lines, "\n")
-
-	// Wrap in panel with gradient border (active since kanban is full-screen)
-	return styles.RenderPanel(content, width, height, true)
-}
-
-// renderKanbanCardLine renders a single line of a kanban card.
-// lineIdx: 0=name, 1=agent, 2=task, 3=stats
-func (p *Plugin) renderKanbanCardLine(wt *Worktree, lineIdx, width int, isSelected bool) string {
-	var content string
-
-	switch lineIdx {
-	case 0:
-		// Line 0: Status icon + name (rune-safe for Unicode)
-		name := wt.Name
-		maxNameLen := width - 3 // Account for icon and space
-		if runes := []rune(name); len(runes) > maxNameLen {
-			name = string(runes[:maxNameLen-3]) + "..."
-		}
-		content = fmt.Sprintf(" %s %s", wt.Status.Icon(), name)
-	case 1:
-		// Line 1: Agent type
-		agentStr := ""
-		if wt.Agent != nil {
-			agentStr = "  " + string(wt.Agent.Type)
-		} else if wt.ChosenAgentType != "" && wt.ChosenAgentType != AgentNone {
-			agentStr = "  " + string(wt.ChosenAgentType)
-		}
-		content = agentStr
-	case 2:
-		// Line 2: Task ID (rune-safe for Unicode)
-		if wt.TaskID != "" {
-			taskStr := wt.TaskID
-			maxLen := width - 2
-			if runes := []rune(taskStr); len(runes) > maxLen {
-				taskStr = string(runes[:maxLen-3]) + "..."
-			}
-			content = "  " + taskStr
-		}
-	case 3:
-		// Line 3: Stats (+/- lines)
-		if wt.Stats != nil && (wt.Stats.Additions > 0 || wt.Stats.Deletions > 0) {
-			content = fmt.Sprintf("  +%d -%d", wt.Stats.Additions, wt.Stats.Deletions)
-		}
-	}
-
-	// Pad to width
-	contentWidth := lipgloss.Width(content)
-	if contentWidth < width {
-		content += strings.Repeat(" ", width-contentWidth)
-	}
-
-	// Apply styling
-	if isSelected {
-		return styles.ListItemSelected.Width(width).Render(content)
-	}
-
-	// Dim non-name lines
-	if lineIdx > 0 {
-		return styles.Muted.Width(width).Render(content)
-	}
-
-	return lipgloss.NewStyle().Width(width).Render(content)
 }
